added hash to base for prime test
