+++ src/arkworks/fermat/mod.rs
@@ -1,5 +1,6 @@
 pub mod constraints;
+pub mod hasher;
 pub mod modpow_circut;
 pub mod modulo;
 pub mod prime_circut;
-pub mod shatry;
+// pub mod shatry;
