+++ src/arkworks/fermat/prime_circut.rs
@@ -21,6 +21,7 @@ use rand::rngs::StdRng;
 use rand::SeedableRng;
 
 use super::constraints::{fermat_circuit, fermat_constructor};
+use super::hasher::hash_to_bytes;
 use super::modulo;
 use itertools::Itertools;
 use sha2::{Digest, Sha256};
@@ -38,58 +39,6 @@ pub struct PrimeCheck<ConstraintF: PrimeField> {
     fermat_circuit: fermat_circuit<ConstraintF>,
 }
 
-fn hash_using_fpvar<ConstraintF: PrimeField>(
-    x_var: FpVar<ConstraintF>,
-    i_var: FpVar<ConstraintF>,
-) -> Vec<u8> {
-    let mut sha256_var = Sha256Gadget::default();
-    // compute x+i:
-    let x_plus_i = x_var + i_var;
-    // convert x_plus_i to bytes:
-    let x_plus_i_bytes = x_plus_i.to_bytes().unwrap();
-    // calculate the hash(x+i):
-    sha256_var.update(&x_plus_i_bytes).unwrap();
-    let calculated_a_i: DigestVar<ConstraintF> = sha256_var.finalize().unwrap();
-    calculated_a_i.value().unwrap().to_vec()
-}
-
-// hash to bytes:
-fn hash_to_bytes<ConstraintF: PrimeField>(x_plus_j: FpVar<ConstraintF>) -> DigestVar<ConstraintF> {
-    let mut sha256_var = Sha256Gadget::default();
-    // convert x_plus_j to bytes:
-    let x_plus_j_bytes = x_plus_j.to_bytes().unwrap();
-    // calculate the hash(x+j):
-    sha256_var.update(&x_plus_j_bytes).unwrap();
-
-    let result = sha256_var.finalize().unwrap();
-    result
-}
-
-fn generate_bases_a<ConstraintF: PrimeField>(
-    cs: ConstraintSystemRef<ConstraintF>,
-    r: FpVar<ConstraintF>,
-) -> Vec<FpVar<ConstraintF>> {
-    let mut a_j_s = vec![];
-    for j in 0..K {
-        let mut sha256_var = Sha256Gadget::default();
-        let r = r.to_bytes().unwrap();
-        let j_bytes = FpVar::<ConstraintF>::constant(ConstraintF::from(j as u64))
-            .to_bytes()
-            .unwrap();
-        sha256_var.update(&r).unwrap();
-        sha256_var.update(&j_bytes).unwrap();
-        let result: DigestVar<ConstraintF> = sha256_var.finalize().unwrap(); // a_i = hash(r || j)
-        let a_j_fpvar =
-            FpVar::<ConstraintF>::new_witness(ark_relations::ns!(cs, "r_j_fpvar"), || {
-                Ok(ConstraintF::from_le_bytes_mod_order(
-                    &result.to_bytes().unwrap().value().unwrap(),
-                ))
-            })
-            .unwrap();
-        a_j_s.push(a_j_fpvar);
-    }
-    a_j_s
-}
 // implement the constraints for the circuit:
 impl<ConstraintF: PrimeField> ConstraintSynthesizer<ConstraintF> for PrimeCheck<ConstraintF> {
     fn generate_constraints(
